Task4and5/requirements.txt
Django==4.2
djangorestframework==3.14.0
channels==4.0.0
channels-redis==4.1.0
djongo==1.3.6
pymongo==4.6.1
